Factor out memory allocation into separate headers
